creating the library itself
yeah, now the library is actually A LIBRARY.
